modify alarm-config domain

+++ jedi-admin/src/main/java/xyz/hellothomas/jedi/admin/infrastructure/mapper/AlarmConfigMapper.java
@@ -19,7 +19,7 @@ public interface AlarmConfigMapper {
 
     List<AlarmConfig> selectByExample(AlarmConfigExample example);
 
-    AlarmConfig selectByPrimaryKey(Integer id);
+    AlarmConfig selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") AlarmConfig record, @Param("example") AlarmConfigExample example);
